refactor: move unsaved agent def back to assistant folder
I did this because the type is one of an assistant not one of an action
