+++ front/lib/actions/types/agent.ts
@@ -42,13 +42,6 @@ import type {
   WebsearchActionRunningEvents,
   WebsearchConfigurationType,
 } from "@app/lib/actions/websearch";
-import type {
-  AgentConfigurationScope,
-  AgentModelConfigurationType,
-  AgentStatus,
-  Editor,
-} from "@app/types";
-import type { TagType } from "@app/types/tag";
 
 /**
  * Agent Action configuration when setting up the agent.
@@ -116,22 +109,6 @@ export type UnsavedServerSideAgentActionConfigurationType = {
     Extract<ServerSideAgentActionConfigurationType, { type: K }>
   >;
 }[ServerSideAgentActionConfigurationType["type"]];
-
-export type UnsavedAgentConfigurationType = {
-  name: string;
-  description: string;
-  instructions: string | null;
-  pictureUrl: string;
-  status: AgentStatus;
-  scope: Exclude<AgentConfigurationScope, "global">;
-  model: AgentModelConfigurationType;
-  actions: UnsavedServerSideAgentActionConfigurationType[];
-  templateId?: string | null;
-  maxStepsPerRun?: number;
-  visualizationEnabled: boolean;
-  tags: TagType[];
-  editors: Editor[];
-};
 
 // Each AgentActionConfigurationType is capable of generating this type at runtime to specify which
 // inputs should be generated by the model. As an example, to run the retrieval action for which the
